TEST/python-app/requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
seaborn
PyJWT